[update] 独立转录逻辑等
1. 修改build.yml与requirements.txt，打包时不再添加torch相关依赖
2. 将保留的pyautogui版的按键发送与MP3转录MIDI转移到tools文件夹内，避免打包时将其包含导致需要引入相关依赖
3. 移除UI中的转炉按钮以及相关逻辑，MP3转录MIDI独立出来，仅供运行代码使用
4. 添加MP3示例文件
5. 修复无法单独运行转录文件的问题
